chore: simplify add block button
