+++ packages/core/flow-engine/src/components/subModel/AddBlockButton.tsx
@@ -14,27 +14,9 @@ import { FlowModel } from '../../models';
 import { AddSubModelButton, AddSubModelButtonProps } from './AddSubModelButton';
 import _ from 'lodash';
 
-export interface AddBlockButtonProps extends Omit<AddSubModelButtonProps, 'subModelType' | 'subModelKey' | 'buildSubModelParams'> {
-  /**
-   * 父模型类名，用于确定支持的块类型
-   */
-  ParentModelClass?: string | typeof FlowModel;
-  
-  /**
-   * 自定义块类型列表，如果不提供则使用默认的块类型
-   */
-  blockTypes?: Array<{
-    key: string;
-    label: string;
-    icon?: React.ReactNode;
-    modelClass: string;
-    [key: string]: any;
-  }>;
-  
-  /**
-   * 子模型在父模型中的键名，默认为 'blocks'
-   */
+interface AddBlockButtonProps extends Omit<AddSubModelButtonProps, 'subModelType' | 'subModelKey'> {
   subModelKey?: string;
+  subModelType?: 'object' | 'array';
 }
 
 /**
@@ -49,59 +31,19 @@ export interface AddBlockButtonProps extends Omit<AddSubModelButtonProps, 'subMo
  * ```
  */
 export const AddBlockButton: React.FC<AddBlockButtonProps> = observer(({
-  model,
   ParentModelClass='BlockFlowModel',
   subModelKey = 'blocks',
   children = 'Add block',
+  subModelType = 'array',
   ...props
 }) => {
-  const blockTypes = useMemo<{
-    key: string;
-    label: string;
-    icon?: React.ReactNode;
-    item: typeof FlowModel;
-  }[]>(() => {
-    const blockClasses = model.flowEngine.filterModelClassByParent(ParentModelClass);
-    const registeredBlocks = [];
-    for (const [className, ModelClass] of blockClasses) {
-      if (ModelClass.meta) {
-        registeredBlocks.push({
-          key: className,
-          label: ModelClass.meta?.title,
-          icon: ModelClass.meta?.icon,
-          item: ModelClass,
-        });
-      }
-    }
-    return registeredBlocks;
-  }, [model, ParentModelClass]);
-  
-  const buildSubModelParams = React.useCallback((info: {
-    key: string;
-  }) => {
-    const blockType = blockTypes.find(type => type.key === info.key);
-    
-    if (!blockType) {
-      throw new Error(`Unknown block type: ${info.key}`);
-    }
-
-    if (blockType.item.meta?.defaultOptions) {
-      return { ..._.cloneDeep(blockType.item.meta?.defaultOptions), use: blockType.key };
-    } else {
-      return {
-        use: blockType.key,
-      }
-    }
-  }, [blockTypes]);
 
   return (
     <AddSubModelButton
       {...props}
-      model={model}
-      subModelType="array"
       subModelKey={subModelKey}
-      items={blockTypes}
-      buildSubModelParams={buildSubModelParams}
+      ParentModelClass={ParentModelClass}
+      subModelType={subModelType}
     >
       {children}
     </AddSubModelButton>
